working csv parse

+++ app/src/main/java/com/example/antidiabet1/item_classes/FoodItem.kt
@@ -2,8 +2,8 @@ package com.example.antidiabet1.item_classes
 
 class FoodItem(
     val name: String,
-    val carbons: Int,
-    val proteins: Int,
-    val fats: Int,
-    val calories: Int
+    val carbons: Double,
+    val proteins: Double,
+    val fats: Double,
+    val calories: Double
 ) { }
